packages/gpu_video_utils/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpu_video_utils"
version = "0.1.0"
description = "GPU-accelerated video processing utilities using PyNvVideoCodec"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "nvidia-pynvvideocodec>=0.0.7",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
